Improved color visualization
Made the optimal path create a gradient, could maybe use a little bit
of work, but for now it's fine

Also found a slight error, Tile,set_color() could attempt to set the
color of a tile about 255, which is the highest value allowed for any
rgb. This would cause pygame to crash. It now clips the value